Restore and update rsx_lib_vulkan
